refactored to separate base classes
