+++ src/gdm/__init__.py
@@ -33,66 +33,72 @@ from gdm.quantities import (
     Irradiance,
 )
 
-from gdm.distribution.components.distribution_substation import (
-    DistributionSubstation,
+from gdm.distribution.components.base.distribution_branch_base import DistributionBranchBase
+from gdm.distribution.components.base.distribution_component_base import DistributionComponentBase
+from gdm.distribution.components.base.distribution_switch_base import DistributionSwitchBase
+from gdm.distribution.components.base.distribution_transformer_base import (
+    DistributionTransformerBase,
 )
-from gdm.distribution.components.distribution_component import (
-    DistributionComponent,
+
+from gdm.distribution.components.distribution_bus import DistributionBus
+from gdm.distribution.components.distribution_capacitor import (
+    DistributionCapacitor,
 )
 from gdm.distribution.components.distribution_feeder import DistributionFeeder
-from gdm.distribution.components.distribution_bus import DistributionBus
-from gdm.distribution.components.distribution_vsource import (
-    PhaseVoltageSourceEquipment,
-    VoltageSourceEquipment,
-    DistributionVoltageSource,
-)
-from gdm.distribution.components.distribution_branch import (
-    DistributionBranch,
-    SwitchedDistributionBranch,
-)
-from gdm.distribution.components.sequence_impedance_branch import (
-    SequenceImpedanceBranch,
-)
-from gdm.distribution.components.matrix_impedance_branch import (
-    MatrixImpedanceBranch,
-)
-from gdm.distribution.components.geometry_branch import GeometryBranch
-from gdm.distribution.components.distribution_switch import DistributionSwitch
-from gdm.distribution.components.matrix_impedance_switch import (
-    MatrixImpedanceSwitch,
-)
-from gdm.distribution.components.distribution_fuse import DistributionFuse
-from gdm.distribution.components.matrix_impedance_fuse import (
-    MatrixImpedanceFuse,
-)
-from gdm.distribution.components.distribution_recloser import (
-    DistributionRecloser,
-)
-from gdm.distribution.components.matrix_impedance_recloser import (
-    MatrixImpedanceRecloser,
-)
 from gdm.distribution.components.distribution_load import DistributionLoad
-from gdm.distribution.components.distribution_transformer import (
-    DistributionTransformer,
-)
 from gdm.distribution.components.distribution_regulator import (
     DistributionRegulator,
 )
 from gdm.distribution.components.distribution_solar import (
     DistributionSolar,
 )
-from gdm.distribution.components.distribution_capacitor import (
-    DistributionCapacitor,
+from gdm.distribution.components.distribution_substation import (
+    DistributionSubstation,
+)
+from gdm.distribution.components.distribution_transformer import (
+    DistributionTransformer,
+)
+from gdm.distribution.components.distribution_vsource import (
+    DistributionVoltageSource,
+)
+from gdm.distribution.components.geometry_branch import GeometryBranch
+from gdm.distribution.components.matrix_impedance_branch import (
+    MatrixImpedanceBranch,
+)
+from gdm.distribution.components.matrix_impedance_fuse import (
+    MatrixImpedanceFuse,
+)
+from gdm.distribution.components.matrix_impedance_recloser import (
+    MatrixImpedanceRecloser,
+)
+from gdm.distribution.components.matrix_impedance_switch import (
+    MatrixImpedanceSwitch,
+)
+from gdm.distribution.components.sequence_impedance_branch import (
+    SequenceImpedanceBranch,
 )
 
-from gdm.distribution.equipment.matrix_impedance_branch_equipment import (
-    MatrixImpedanceBranchEquipment,
+from gdm.distribution.equipment.base.matrix_impedance_branch_equipment_base import (
+    MatrixImpedanceBranchEquipmentBase,
 )
 from gdm.distribution.equipment.bare_conductor_equipment import (
     BareConductorEquipment,
 )
+from gdm.distribution.equipment.capacitor_equipment import CapacitorEquipment
 from gdm.distribution.equipment.concentric_cable_equipment import (
     ConcentricCableEquipment,
+)
+from gdm.distribution.equipment.distribution_transformer_equipment import (
+    DistributionTransformerEquipment,
+    WindingEquipment,
+)
+from gdm.distribution.equipment.geometry_branch_equipment import (
+    GeometryBranchEquipment,
+)
+from gdm.distribution.equipment.inverter_equipment import InverterEquipment
+from gdm.distribution.equipment.load_equipment import LoadEquipment
+from gdm.distribution.equipment.matrix_impedance_branch_equipment import (
+    MatrixImpedanceBranchEquipment,
 )
 from gdm.distribution.equipment.matrix_impedance_fuse_equipment import (
     MatrixImpedanceFuseEquipment,
@@ -103,43 +109,38 @@ from gdm.distribution.equipment.matrix_impedance_recloser_equipment import (
 from gdm.distribution.equipment.matrix_impedance_switch_equipment import (
     MatrixImpedanceSwitchEquipment,
 )
-from gdm.distribution.equipment.recloser_controller_equipment import (
-    RecloserControllerEquipment,
-)
 from gdm.distribution.equipment.phase_capacitor_equipment import (
     PhaseCapacitorEquipment,
 )
-from gdm.distribution.equipment.capacitor_equipment import CapacitorEquipment
-from gdm.distribution.equipment.geometry_branch_equipment import (
-    GeometryBranchEquipment,
+from gdm.distribution.equipment.phase_load_equipment import PhaseLoadEquipment
+
+from gdm.distribution.equipment.recloser_controller_equipment import (
+    RecloserControllerEquipment,
 )
 from gdm.distribution.equipment.sequence_impedance_branch_equipment import (
     SequenceImpedanceBranchEquipment,
 )
-from gdm.distribution.equipment.phase_load_equipment import PhaseLoadEquipment
-from gdm.distribution.equipment.load_equipment import LoadEquipment
-
-from gdm.distribution.equipment.distribution_transformer_equipment import (
-    DistributionTransformerEquipment,
-    WindingEquipment,
-)
 from gdm.distribution.equipment.solar_equipment import SolarEquipment
-from gdm.distribution.equipment.inverter_equipment import InverterEquipment
+from gdm.distribution.equipment.voltagesource_equipment import VoltageSourceEquipment
 
-from gdm.distribution.controllers.distribution_inverter_controller import (
-    InverterController,
-    PowerfactorInverterController,
-    VoltVarInverterController,
-    VoltVarVoltWattInverterController,
-    VoltWattInverterController,
-)
+
+from gdm.distribution.controllers.base.capacitor_controller_base import CapacitorControllerBase
+from gdm.distribution.controllers.base.inverter_controller_base import InverterControllerBase
 from gdm.distribution.controllers.distribution_capacitor_controller import (
-    CapacitorController,
     VoltageCapacitorController,
     ActivePowerCapacitorController,
     ReactivePowerCapacitorController,
     CurrentCapacitorController,
     DailyTimedCapacitorController,
+)
+from gdm.distribution.controllers.distribution_inverter_controller import (
+    PowerfactorInverterController,
+    VoltVarInverterController,
+    VoltVarVoltWattInverterController,
+    VoltWattInverterController,
+)
+from gdm.distribution.controllers.distribution_recloser_controller import (
+    DistributionRecloserController,
 )
 from gdm.distribution.controllers.distribution_regulator_controller import (
     RegulatorController,
@@ -147,9 +148,7 @@ from gdm.distribution.controllers.distribution_regulator_controller import (
 from gdm.distribution.controllers.distribution_switch_controller import (
     DistributionSwitchController,
 )
-from gdm.distribution.controllers.distribution_recloser_controller import (
-    DistributionRecloserController,
-)
+
 
 from gdm.distribution.sequence_pair import SequencePair
 from gdm.distribution.limitset import ThermalLimitSet, VoltageLimitSet
