+++ src/gdm/distribution/components/base/distribution_component_base.py
@@ -2,6 +2,7 @@
 
 from typing_extensions import Annotated
 from typing import Optional
+from abc import ABC
 
 from pydantic import Field
 from infrasys import Component
@@ -10,18 +11,10 @@ from gdm.distribution.components.distribution_feeder import DistributionFeeder
 from gdm.distribution.components.distribution_substation import DistributionSubstation
 
 
-class DistributionComponent(Component):
+class DistributionComponentBase(Component, ABC):
     """Interface for simple distribution component."""
 
     substation: Annotated[
         Optional[DistributionSubstation], Field(None, description="Name of the substation.")
     ]
     feeder: Annotated[Optional[DistributionFeeder], Field(None, description="Name of the feeder.")]
-
-    @classmethod
-    def example(cls) -> "DistributionComponent":
-        return DistributionComponent(
-            name="BaseComponent",
-            substation=DistributionSubstation.example(),
-            feeder=DistributionFeeder.example()
-        )
