+++ src/gdm/distribution/components/geometry_branch.py
@@ -4,17 +4,27 @@ from typing import Annotated
 
 from pydantic import Field
 
-from gdm.distribution.components.distribution_branch import DistributionBranch
-from gdm.distribution.equipment.geometry_branch_equipment import GeometryBranchEquipment
+from gdm.distribution.components.base.distribution_branch_base import (
+    DistributionBranchBase,
+)
+from gdm.distribution.equipment.geometry_branch_equipment import (
+    GeometryBranchEquipment,
+)
 from gdm.distribution.components.distribution_feeder import DistributionFeeder
-from gdm.distribution.components.distribution_substation import DistributionSubstation
+from gdm.distribution.components.distribution_substation import (
+    DistributionSubstation,
+)
+from gdm.distribution.components.distribution_bus import DistributionBus
+from gdm.distribution.distribution_enum import Phase
+from gdm.quantities import PositiveVoltage, PositiveDistance
 
 
-class GeometryBranch(DistributionBranch):
+class GeometryBranch(DistributionBranchBase):
     """Interface for geometry based lines."""
 
     equipment: Annotated[
-        GeometryBranchEquipment, Field(..., description="Geometry branch equipment.")
+        GeometryBranchEquipment,
+        Field(..., description="Geometry branch equipment."),
     ]
 
     def validate_fields(self) -> "GeometryBranch":
@@ -27,13 +37,28 @@ class GeometryBranch(DistributionBranch):
     @classmethod
     def example(cls) -> "GeometryBranch":
         """Example for geometry branch."""
-        base_branch = DistributionBranch.example()
-        return GeometryBranch(
-            buses=base_branch.buses,
-            length=base_branch.length,
-            phases=base_branch.phases,
-            name=base_branch.name,
+        bus1 = DistributionBus(
+            voltage_type="line-to-ground",
+            phases=[Phase.A, Phase.B, Phase.C],
+            nominal_voltage=PositiveVoltage(400, "volt"),
             substation=DistributionSubstation.example(),
             feeder=DistributionFeeder.example(),
+            name="Branch-DistBus1",
+        )
+        bus2 = DistributionBus(
+            voltage_type="line-to-ground",
+            phases=[Phase.A, Phase.B, Phase.C],
+            nominal_voltage=PositiveVoltage(400, "volt"),
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
+            name="Branch-DistBus2",
+        )
+        return GeometryBranch(
+            buses=[bus1, bus2],
+            length=PositiveDistance(130.2, "meter"),
+            phases=[Phase.A, Phase.B, Phase.C],
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
+            name="DistBranch1",
             equipment=GeometryBranchEquipment.example(),
         )
