+++ src/gdm/distribution/distribution_graph.py
@@ -4,8 +4,8 @@ import networkx as nx
 from infrasys import System
 
 from gdm import (
-    DistributionTransformer,
-    DistributionBranch,
+    DistributionTransformerBase,
+    DistributionBranchBase,
     DistributionBus,
 )
 
@@ -18,9 +18,9 @@ def build_graph_from_system(system: System) -> nx.Graph:
     for node in system.get_components(DistributionBus):
         graph.add_node(node.name)
 
-    edges: list[DistributionBranch | DistributionTransformer] = list(
-        system.get_components(DistributionBranch)
-    ) + list(system.get_components(DistributionTransformer))
+    edges: list[DistributionBranchBase | DistributionTransformerBase] = list(
+        system.get_components(DistributionBranchBase)
+    ) + list(system.get_components(DistributionTransformerBase))
 
     for edge in edges:
         graph.add_edge(
