+++ tests/test_distribution_examples.py
@@ -4,8 +4,6 @@ import pytest
 
 from gdm import (
     DistributionBus,
-    DistributionComponent,
-    PhaseVoltageSourceEquipment,
     VoltageSourceEquipment,
     DistributionVoltageSource,
     ConcentricCableEquipment,
@@ -13,7 +11,6 @@ from gdm import (
     SequenceImpedanceBranchEquipment,
     MatrixImpedanceBranchEquipment,
     GeometryBranchEquipment,
-    DistributionBranch,
     MatrixImpedanceBranch,
     SequenceImpedanceBranch,
     GeometryBranch,
@@ -29,26 +26,20 @@ from gdm import (
     SolarEquipment,
     DistributionSolar,
     DistributionRegulator,
-    CapacitorController,
     VoltageCapacitorController,
     ActivePowerCapacitorController,
     ReactivePowerCapacitorController,
     CurrentCapacitorController,
     DailyTimedCapacitorController,
-    InverterController,
     PowerfactorInverterController,
     VoltVarInverterController,
     VoltVarVoltWattInverterController,
     RegulatorController,
-    DistributionSwitch,
-    DistributionFuse,
-    SwitchedDistributionBranch,
     MatrixImpedanceFuseEquipment,
     MatrixImpedanceFuse,
     MatrixImpedanceSwitch,
     Curve,
     TimeCurrentCurve,
-    DistributionRecloser,
     MatrixImpedanceRecloserEquipment,
     MatrixImpedanceRecloser,
     RecloserControllerEquipment,
@@ -61,8 +52,6 @@ from gdm import (
 
 DIST_INTERFACES = [
     DistributionBus,
-    DistributionComponent,
-    PhaseVoltageSourceEquipment,
     VoltageSourceEquipment,
     DistributionVoltageSource,
     ConcentricCableEquipment,
@@ -70,7 +59,6 @@ DIST_INTERFACES = [
     SequenceImpedanceBranchEquipment,
     MatrixImpedanceBranchEquipment,
     GeometryBranchEquipment,
-    DistributionBranch,
     MatrixImpedanceBranch,
     SequenceImpedanceBranch,
     GeometryBranch,
@@ -88,26 +76,20 @@ DIST_INTERFACES = [
     DistributionRegulator,
     DistributionSolar,
     DistributionRegulator,
-    CapacitorController,
     VoltageCapacitorController,
     ActivePowerCapacitorController,
     ReactivePowerCapacitorController,
     CurrentCapacitorController,
     DailyTimedCapacitorController,
     Curve,
-    InverterController,
     PowerfactorInverterController,
     VoltVarInverterController,
     VoltVarVoltWattInverterController,
     RegulatorController,
-    DistributionSwitch,
-    DistributionFuse,
-    SwitchedDistributionBranch,
     MatrixImpedanceFuseEquipment,
     MatrixImpedanceFuse,
     MatrixImpedanceSwitch,
     TimeCurrentCurve,
-    DistributionRecloser,
     MatrixImpedanceRecloserEquipment,
     MatrixImpedanceRecloser,
     RecloserControllerEquipment,
