+++ tests/test_json_schema.py
@@ -2,8 +2,6 @@ import pytest
 
 from gdm import (
     DistributionBus,
-    DistributionComponent,
-    PhaseVoltageSourceEquipment,
     VoltageSourceEquipment,
     DistributionVoltageSource,
     ConcentricCableEquipment,
@@ -11,7 +9,6 @@ from gdm import (
     SequenceImpedanceBranchEquipment,
     MatrixImpedanceBranchEquipment,
     GeometryBranchEquipment,
-    DistributionBranch,
     MatrixImpedanceBranch,
     SequenceImpedanceBranch,
     GeometryBranch,
@@ -27,26 +24,20 @@ from gdm import (
     SolarEquipment,
     DistributionSolar,
     DistributionRegulator,
-    CapacitorController,
     VoltageCapacitorController,
     ActivePowerCapacitorController,
     ReactivePowerCapacitorController,
     CurrentCapacitorController,
     DailyTimedCapacitorController,
-    InverterController,
     PowerfactorInverterController,
     VoltVarInverterController,
     VoltVarVoltWattInverterController,
     RegulatorController,
-    DistributionSwitch,
-    DistributionFuse,
-    SwitchedDistributionBranch,
     MatrixImpedanceFuseEquipment,
     MatrixImpedanceFuse,
     MatrixImpedanceSwitch,
     Curve,
     TimeCurrentCurve,
-    DistributionRecloser,
     MatrixImpedanceRecloserEquipment,
     MatrixImpedanceRecloser,
     RecloserControllerEquipment,
@@ -60,8 +51,6 @@ from gdm import (
 
 CLASSES_TO_CHECK_JSON_SCHEMA = [
     DistributionBus,
-    DistributionComponent,
-    PhaseVoltageSourceEquipment,
     VoltageSourceEquipment,
     DistributionVoltageSource,
     ConcentricCableEquipment,
@@ -69,7 +58,6 @@ CLASSES_TO_CHECK_JSON_SCHEMA = [
     SequenceImpedanceBranchEquipment,
     MatrixImpedanceBranchEquipment,
     GeometryBranchEquipment,
-    DistributionBranch,
     MatrixImpedanceBranch,
     SequenceImpedanceBranch,
     GeometryBranch,
@@ -85,26 +73,20 @@ CLASSES_TO_CHECK_JSON_SCHEMA = [
     SolarEquipment,
     DistributionSolar,
     DistributionRegulator,
-    CapacitorController,
     VoltageCapacitorController,
     ActivePowerCapacitorController,
     ReactivePowerCapacitorController,
     CurrentCapacitorController,
     DailyTimedCapacitorController,
-    InverterController,
     PowerfactorInverterController,
     VoltVarInverterController,
     VoltVarVoltWattInverterController,
     RegulatorController,
-    DistributionSwitch,
-    DistributionFuse,
-    SwitchedDistributionBranch,
     MatrixImpedanceFuseEquipment,
     MatrixImpedanceFuse,
     MatrixImpedanceSwitch,
     Curve,
     TimeCurrentCurve,
-    DistributionRecloser,
     MatrixImpedanceRecloserEquipment,
     MatrixImpedanceRecloser,
     RecloserControllerEquipment,
